+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.0.5
+
+- Added new component `<skawa-grid>`
+
+
+## 0.0.4
+
+- CKEditor accepts initial value as `content` input property
+
+## 0.0.3+1
+
+- Fixed transformer usage
+
+## 0.0.3
+
+- Added new component `<skawa-ckeditor>`
+
 ## 0.0.2
 
 - Added tests with travis integration